requirements.txt
amqp==2.5.1
asn1crypto==0.24.0
beautifulsoup4==4.8.0
billiard==3.6.1.0
bs4==0.0.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
exponent-server-sdk==0.2.0
firebase==3.0.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.0
Flask-SSLify==0.1.5
gcloud==0.18.3
googleapis-common-protos==1.6.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
google-cloud-vision==0.41.0
google-cloud==0.34.0
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jwcrypto==0.6.0
kombu==4.6.4
MarkupSafe==1.1.1
more-itertools==7.2.0
Naked==0.1.31
oauth2client==4.1.3
progress==1.5
protobuf==3.9.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodome==3.8.2
pyfcm==1.4.7
PyJWT==1.7.1
python-jwt==3.2.4
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
requests-toolbelt==0.9.1
rq==1.1.0
rsa==4.0
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.2
sseclient==0.0.24
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.5
zipp==0.6.0
